requirements.txt
certifi==2019.6.16
chardet==3.0.4
Deprecated==1.2.5
idna==2.8
PyGithub==1.43.7
PyJWT==1.7.1
requests==2.22.0
schedule==0.6.0
urllib3==1.25.3
wrapt==1.11.2
